Remove PSBool to ledger API
